requirements.txt
streamlit
openai-whisper
srt
moviepy
torch
numpy
setuptools-rust
google-generativeai